refactor: move bench diagnostic + phase-event types to homeboy-extension-contract
Stage-2 cut continuing the bench subsystem breakdown. Splits the pure
diagnostic and phase-event types out of their behavior:

- diagnostic.rs: BenchDiagnostic, BenchDiagnosticSource (the collect_diagnostics
  function, which consumes BenchResults, stays in core)
- phase_events.rs: BenchPhaseEvent, BenchPhaseSummary,
  BenchPhaseFailureClassification (evaluate_phase_events stays in core)

All five are pure serde types referencing only each other + BTreeMap. Moved
their is_zero_usize serde helper too. Core re-exports them (pub use) so the
bench module paths stay stable for ~18 consumers.

These types are field dependencies of the still-entangled BenchResults/
BenchScenario cluster; moving them (plus the observation::timeline relocation)
continues clearing the path for those larger result models.
