+++ crates/homeboy-extension-contract/src/lib.rs
@@ -11,8 +11,13 @@
 
 pub mod action_types;
 pub mod autofix_config;
+pub mod bench_diagnostics;
 pub mod bench_result;
 pub mod capability;
+pub use bench_diagnostics::{
+    BenchDiagnostic, BenchDiagnosticSource, BenchPhaseEvent, BenchPhaseFailureClassification,
+    BenchPhaseSummary,
+};
 pub use bench_result::{
     BenchChildCommandFailure, BenchMemory, BenchMetricDirection, BenchMetricPhase,
     BenchMetricPolicy, BenchMetrics, BenchProvenance, BenchProvenanceLink, BenchRunExecution,
